Crate atm

atm-cli is a command line tool for generating and working with MIDI files. It was purpose-built for All the Music, LLC to assist in its mission to enable musicians to make all of their music without the fear of frivolous copyright lawsuits. All code is released into the public domain via the Creative Commons Attribute 4.0 International License. If you're looking for a Rust library to generate and work with MIDI files, check out the libatm project, on which this tool relies. For more information on All the Music, check out allthemusic.info.

Structs

BatchedMIDIArchive
Container for tar archive of MIDI files

Enums

BatchedMIDIArchiveState
State of a BatchedMIDIArchive

Functions

gen_partition_size
Calculate number of MIDI notes per partition.

gen_path
Generate a path representing the partition for the given MIDI hash separated by the OS path separator.

gen_sequences
Generate all permutations (with replacement) of given length from the given sequence of MIDI notes.