Module atm::storage

MIDI file storage backends

Re-exports

pub use batch_tar_file::BatchTarFile;

pub use tar_file::TarFile;

pub use tar_gz_file::TarGzFile;

Modules

batch_tar_file
Batch archive storage backend

tar_file
Tar archive storage backend

tar_gz_file
Gzip-compressed Tar archive storage backend

Structs

MIDIHashPathGenerator
Path generator that produces the hash of a MIDI file as the filename without any parent directories (see: MIDIFile::gen_hash). This path generator is useful for smaller datasets.

PartitionPathGenerator
Path generator for storage backends that support partitioned output schemes

TarArchive
Tar archive storage backend. Used by other storage backends as the underlying storage object.

Enums

PartitionPathGeneratorError
Error type for PartitionPathGenerator

PathGeneratorError
Error type for PathGenerator

StorageState
Whether storage backend is open, or is closed for writing

TarArchiveError
Error type for TarArchive (wrapping around std::io::Error and PathGeneratorError)

Traits

IntoInner
Trait to implement functionality for storage backends to expose the underlying (inner) storage object.

PathGenerator
Trait to generate storage path for MIDI files in storage backends

StorageBackend
Trait to implement storage backends for MIDI files generated from libatm::MIDIFile.