Trait atm::storage::IntoInner

pub trait IntoInner: StorageBackend {
    type Inner;
    fn into_inner(self) -> Result<Self::Inner, Self::Error>;
}

Trait to implement functionality for storage backends to expose the underlying (inner) storage object.

Associated Types

type Inner

Type of inner object

Loading content...
Required methods

fn into_inner(self) -> Result<Self::Inner, Self::Error>

Finish writing storage backends and return the inner object

Loading content...