Crate atm

atm-cli is a command line tool for generating and working with MIDI files. It was purpose-built for All the Music, LLC to assist in its mission to enable musicians to make all of their music without the fear of frivolous copyright lawsuits. All code is freely available under the Creative Commons Attribution 4.0 International License. If you're looking for a Rust library to generate and work with MIDI files, check out the libatm project, on which this tool relies. For more information on All the Music, check out allthemusic.info.

Modules

storage
MIDI file storage backends

utils
Utilities for generating melodies